Move TreeChevron to Tree.tsx
