+++ packages/kiwi-react/src/bricks/Icon.css
@@ -31,13 +31,3 @@
 		}
 	}
 }
-
-.🥝-tree-chevron {
-	@layer base {
-		rotate: var(--🥝tree-chevron-rotate);
-
-		@media (prefers-reduced-motion: no-preference) {
-			transition: rotate 150ms ease-in-out;
-		}
-	}
-}
