+++ packages/kiwi-react/src/bricks/Icon.tsx
@@ -98,31 +98,3 @@ export const DisclosureArrow = React.forwardRef<
 		/>
 	);
 });
-
-// ----------------------------------------------------------------------------
-
-interface TreeChevronProps extends Omit<Ariakit.RoleProps<"svg">, "children"> {}
-
-export const TreeChevron = React.forwardRef<
-	React.ElementRef<"svg">,
-	TreeChevronProps
->((props, forwardedRef) => {
-	return (
-		<Icon
-			{...props}
-			render={
-				<Ariakit.Role.svg
-					width="16"
-					height="16"
-					fill="currentColor"
-					viewBox="0 0 16 16"
-					render={props.render}
-				>
-					<path d="M4.146 6.146a.5.5 0 0 1 .708 0L8 9.293l3.146-3.147a.5.5 0 0 1 .708.708l-3.5 3.5a.5.5 0 0 1-.708 0l-3.5-3.5a.5.5 0 0 1 0-.708Z" />
-				</Ariakit.Role.svg>
-			}
-			className={cx("🥝-tree-chevron", props.className)}
-			ref={forwardedRef}
-		/>
-	);
-});
